test: add test case for finding log error patterns (#128)
* Add log lines when creating Sift investigation

Should help debug issues when running Sift tools.

* Include Grafana client in cloud test context

Using the refactored NewGrafanaClient function from
PR #127.

* Fetch multiple investigations in tests, in case the first one has no analyses

This can happen if, for example, an investigation was run with a scope
that didn't find any logs, and so no analyses were created.

* Add test for finding error pattern logs

This relies on a new Loki datasource being added to the mcptests
instance, which points to our dev logs. We probably want to get rid of
this and use different logs; I'll look into how we can do that.