Revert back to using separate SSE/HTTP funcs
The mcp-go package decided to revert back to not having unified options
for SSE and streaming HTTP servers, so we need to switch back to using the
old (now no-longer deprecated) functions to set them. Annoyingly the types
for the context functions are distinct but the underlying function is actually
the same, so I added a custom httpContextFunc type which unifies them on our
side so we can reuse the same functions for both.

Also, fix types in tests.