fix: correctly populate 'version' of MCP server
This uses the `runtime/debug` package to get the version of the binary
at build time, which in turn uses git information to populate the
version string.

When running from source using `go run` the version is set to
`(devel)`.

When installed using either `go install` or `go build` the version is
populated like so:

```sh
❯ ./dist/mcp-grafana -version
v0.5.1-0.20250624101014-217ec1a1e5ee
```