Add HTTP header passthrough
Background: we have Grafana behind a reverse proxy that authenticates the
user to a given LDAP based on an authentication ticket (set in the header as `Authorization: Bearer <token>`).

When running mcp-grafana centrally (to maintain user access) we will be
having MCP clients authenticate to obtain their token that will be set
as a header with every request to mcp-grafana. mcp-grafana can use this
same authorization header to authenticate against Grafana for the provided
client.