chore: add server.json for MCP registry
Also update the 'docker' CI workflow to publish to the registry after
building and pushing the image.

This is split out of #291 since it's actually a different thing; this PR
adds our MCP server to the registry, using the Docker image. The other
will add an MCPB manifest which we _could_ use for the registry instead,
and which can be used for Claude Desktop.